Background Hex finished

+++ CatanClient/UI/Texture.cs
@@ -9,13 +9,13 @@ namespace CatanClient.UI
 {
     abstract class Texture
     {
-        public int X { protected set; get; }
-        public int Y { protected set; get; }
-        public int Height { protected set; get; }
-        public int Width { protected set; get; }
+        public float X { protected set; get; }
+        public float Y { protected set; get; }
+        public float Height { protected set; get; }
+        public float Width { protected set; get; }
 
 
-        public Texture(int x, int y,int height, int width)
+        public Texture(float x, float y,float height, float width)
         {
             this.X = x;
             this.Y = y;
